Linker: passes 1 and 2 done. Pass 3 still in development.

+++ toolchain/objdump/stdobj.h
@@ -64,9 +64,9 @@ typedef struct _obj_image_t {
 typedef union {
     uint64_t    u64;
     struct {
-        unsigned n_addr : 15;   // address
-        unsigned n_type : 9;    // symbol type
-        unsigned n_ref  : 24;   // reference to another symbol,
+        uint64_t n_addr : 15;   // address
+        uint64_t n_type : 9;    // symbol type
+        uint64_t n_ref  : 24;   // reference to another symbol,
     } f;                        // or short name
 } nlist_t;
